add files: change_notify.[hc]

+++ src/common/fdir_types.h
@@ -112,7 +112,7 @@ typedef struct fdir_dentry_info {
 } FDIRDEntryInfo;
 
 typedef union {
-    int64_t flags;
+    int32_t flags;
     struct {
         union {
             int flags: 4;
